dockerfile-requirements.txt
click==8.0.1
python-dotenv==0.19.1
pandas==1.1.5
gdown==4.4.0
numpy==1.19.5
matplotlib==3.3.4
setuptools==59.6.0
tensorflow_addons==0.14.0
tqdm
requests
PyYAML
